requirements.txt
# -------------------------------
# Core Framework
# -------------------------------
Flask==3.1.1
flask-cors==6.0.1
Werkzeug==3.1.3
gunicorn==21.2.0

# -------------------------------
# Firebase & Google Cloud
# -------------------------------
firebase-admin==6.9.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.7.2

# -------------------------------
# Gemini (LLM)
# -------------------------------
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15

# -------------------------------
# Environment & Security
# -------------------------------
python-dotenv==1.1.1
cryptography==45.0.5
bcrypt==4.3.0
PyJWT==2.10.1

# -------------------------------
# File Parsing (Resume / JD)
# -------------------------------
python-docx==1.2.0
docx2txt==0.9
PyPDF2==3.0.1
pdfminer.six==20231228
lxml==6.0.0
Pillow==11.3.0
PyMuPDF==1.24.0

# -------------------------------
# NLP (Lightweight)
# -------------------------------
nltk==3.9.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

# -------------------------------
# Data / Utils
# -------------------------------
numpy==2.3.1
scikit-learn==1.7.0
joblib==1.5.1
regex==2024.11.6
requests==2.32.4
tqdm==4.67.1
rich==14.0.0

# -------------------------------
# PDF / Report Generation
# -------------------------------
reportlab==4.2.0
xlsxwriter==3.2.5

# -------------------------------
# Misc (Let pip resolve protobuf automatically)
# -------------------------------
typing_extensions==4.14.0
packaging==25.0
urllib3==2.5.0
certifi==2025.6.15